lib: download-raw: Add md5 hash to the raw_data to enable quicker skip
We sometimes see duplicated data which causes us to have to re-validate
and normalise it when in fact the data hasn't changed. Handle this by
hashing the data locally and storing this in the DB and checking this
value before inserting or updating the existing data. All the data
should be unique because of the time and state information in the
data itself.